backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
pydantic
langchain
langchain-community
langchain-openai
python-dotenv
